requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0
ruff>=0.4.0
